requirements.txt
requests
python-dotenv
Werkzeug
PyYAML
openai
autogen-agentchat
autogen-ext[azure,openai]
boto3
fastapi
uvicorn
pdfplumber
azure-cosmos
python-multipart
jinja2
gunicorn
scipy
hnswlib